refactor: move BookingForm to own folder under container
